feat(server): run webpack HMR as express middleware
Current dev environments run both expressjs and webpack-dev-server,
which leads to confusion over which port to use when looking at the
web UI or visiting a shortlink. It is also inconsistent with how
the live environment is actually set up.

Run webpack as part of expressjs so that we keep to one server.

[deps]
- replace webpack-dev-server with middleware equivalents
- add supporting cast of types
- consolidate to running just `docker-dev` for dev docker cluster

[server]
- create a new webpack-middleware module, deliberately getting eslint
  to overlook the fact that we use dev-only packages, since we do not
  expect to run this in the live environment
- at runtime, get the server to choose between webpack or static,
  depending on `NODE_ENV`

[docker]
- remove references to webpack-dev-server
- add more memory for Node.js to accommodate both express and webpack